fix(en/uniquestream): Fix cloudflare errors and move to multisrc
