+++ multisrc/overrides/dooplay/uniquestream/res/mipmap-hdpi/ic_launcher.png
