+++ multisrc/overrides/dooplay/uniquestream/res/mipmap-mdpi/ic_launcher.png
