+++ multisrc/overrides/dooplay/uniquestream/res/mipmap-xhdpi/ic_launcher.png
