+++ multisrc/overrides/dooplay/uniquestream/res/mipmap-xxhdpi/ic_launcher.png
