+++ multisrc/overrides/dooplay/uniquestream/res/mipmap-xxxhdpi/ic_launcher.png
