--- src/en/uniquestream/build.gradle
@@ -1,13 +1,0 @@
-apply plugin: 'com.android.application'
-apply plugin: 'kotlin-android'
-apply plugin: 'kotlinx-serialization'
-
-ext {
-    extName = 'UniqueStream'
-    pkgNameSuffix = 'en.uniquestream'
-    extClass = '.UniqueStream'
-    extVersionCode = 1
-    libVersion = '13'
-}
-
-apply from: "$rootDir/common.gradle"
